Enable Native Image Layers on Darwin
