+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/c/codegen/CCompilerInvoker.java
@@ -289,7 +289,7 @@ public abstract class CCompilerInvoker {
 
     }
 
-    @SingletonTraits(access = BuildtimeAccessOnly.class, layeredCallbacks = NoLayeredCallbacks.class, other = Disallowed.class)
+    @SingletonTraits(access = BuildtimeAccessOnly.class, layeredCallbacks = NoLayeredCallbacks.class)
     private static class DarwinCCompilerInvoker extends CCompilerInvoker {
 
         DarwinCCompilerInvoker(Path tempDirectory) {
